+++ database/migrations/2021_11_30_181336_create_outgoing_items_table.php
@@ -18,7 +18,7 @@ class CreateOutgoingItemsTable extends Migration
             $table->foreignId('outgoing_id');
             $table->string('item_name');
             $table->integer('quantity')->nullable();
-            $table->integer('cubic_metre')->nullable();
+            $table->decimal('cubic_metre')->nullable();
             $table->timestamps();
 
             $table->foreign('outgoing_id')
